requirements.txt
sentence-transformers
scikit-learn
google-genai
python-dotenv
chromadb
fastapi
uvicorn
beautifulsoup4
requests